requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
black==24.10.0
certifi==2024.8.30
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
flake8==7.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isort==5.13.2
jiter==0.7.0
mccabe==0.7.0
mypy-extensions==1.0.0
openai==1.53.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
python-dotenv==1.0.1
sniffio==1.3.1
tomli==2.0.2
tqdm==4.66.6
typing_extensions==4.12.2
